+++ src/engine/index.ts
@@ -2,5 +2,5 @@ export { OrbitCamera } from './OrbitCamera';
 export { Ray } from './Ray';
 export { GizmoRenderer } from './GizmoRenderer';
 export type { GizmoMode } from './GizmoRenderer';
-export { GizmoInteraction } from './GizmoInteraction';
-export type { GizmoAxis } from './GizmoInteraction';
+export { GizmoInteraction, TRANSLATE_SNAP, ROTATE_SNAP_DEG, SCALE_SNAP } from './GizmoInteraction';
+export type { GizmoAxis, GizmoHandle } from './GizmoInteraction';